Refactor project for testing ease.
