+++ src/pugbot/commands/add.rs
