+++ src/pugbot/commands/remove.rs
